Fix overflow in cache expiration threshold calculation
Agent-Logs-Url: https://github.com/r-buc/mdnsd/sessions/8e84479e-c414-41e6-b994-dfaab8b054e7

+++ cache.c
@@ -69,7 +69,7 @@ cache_service_free(struct cache_service *s)
 static int
 cache_is_expired(time_t t, uint32_t ttl, int frac)
 {
-	if (monotonic_time() - t >= ttl * frac / 100)
+	if (monotonic_time() - t >= (uint64_t)ttl * frac / 100)
 		return 1;
 
 	return 0;
